chore: move `generate_abi` to `cargo_near_build::near::abi::generate::procedure`
