+++ cargo-near-build/Cargo.toml
@@ -18,3 +18,4 @@ libloading = "0.8.3"
 serde_json = "1.0"
 near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
 zstd = "0.13"
+schemars = "0.8"
